+++ app/(setup-layout)/first-steps/page.tsx
@@ -20,7 +20,7 @@ import { useAuthUser } from '@/hooks/auth/useAuthUser'
 import { v4 as uuidv4 } from 'uuid'
 import { routes } from '@/utils/routes'
 import { useApiKeyEncryption } from '@/app/(setup-layout)/first-steps/hooks/useApiKeyEncryption'
-import sellingOptions from '@/app/(without-navbar)/first-steps/data/selling-options.json'
+import sellingOptions from '@/app/(setup-layout)/first-steps/data/selling-options.json'
 
 export default function FirstStepsPage() {
   const [step, setStep] = useState(1)
