+++ utils/ConfigureAmplify.tsx
@@ -1,7 +1,6 @@
 'use client'
 
 import { Amplify } from 'aws-amplify'
-
 import outputs from '@/amplify_outputs.json'
 
 Amplify.configure(outputs, { ssr: true })
